requirements.txt
pandas==2.1.0
numpy<2
kafka-python==2.0.2
pymongo==4.5.0
